move src reader and sink writer to util.io
