+++ components/server/src/main/java/ome/smuggler/config/wiring/crypto/SerializationFactory.java
@@ -5,12 +5,13 @@ import static java.util.Objects.requireNonNull;
 import java.io.InputStream;
 import java.io.OutputStream;
 
-import ome.smuggler.core.convert.SinkWriter;
-import ome.smuggler.core.convert.SourceReader;
 import ome.smuggler.core.io.crypto.*;
 import ome.smuggler.core.types.CryptoConfigSource;
 import ome.smuggler.providers.json.JsonInputStreamReader;
 import ome.smuggler.providers.json.JsonOutputStreamWriter;
+
+import util.io.SinkWriter;
+import util.io.SourceReader;
 
 /**
  * Builds serializers depending on the current crypto configuration.
