+++ components/server/src/main/java/ome/smuggler/providers/json/JsonStreamWriter.java
@@ -5,7 +5,7 @@ import static java.util.Objects.requireNonNull;
 import java.io.Writer;
 import java.util.function.Function;
 
-import ome.smuggler.core.convert.SinkWriter;
+import util.io.SinkWriter;
 
 
 /**
