+++ components/server/src/test/java/ome/smuggler/providers/q/DequeueTaskTest.java
@@ -4,7 +4,7 @@ import static org.hamcrest.Matchers.*;
 import static org.junit.Assert.*;
 import static org.mockito.Mockito.*;
 
-import ome.smuggler.core.convert.SourceReader;
+import util.io.SourceReader;
 import org.apache.activemq.artemis.api.core.ActiveMQBuffer;
 import org.apache.activemq.artemis.api.core.ActiveMQException;
 import org.apache.activemq.artemis.api.core.client.ClientConsumer;
